docker_deploy/requirements.txt
python-telegram-bot
aiohttp
aiosqlite
openpyxl
pandas
fastapi
uvicorn
python-multipart
streamlit
python-docx
Pillow
httpx
python-dotenv
nest_asyncio
APScheduler
